feat: :sparkles: Add cupertino delegate support
