+++ lib/flutter_kurdish_localization.dart
@@ -1,4 +1,6 @@
 library flutter_kurdish_localization;
 
+export 'package:flutter_kurdish_localization/kurdish_cupertino_localization_delegate.dart';
 export 'package:flutter_kurdish_localization/kurdish_material_localization_delegate.dart';
 export 'package:flutter_kurdish_localization/kurdish_widget_localization_delegate.dart';
+export 'package:flutter_kurdish_localization/kurdish_localization_delegates.dart';
