+++ pubspec.yaml
@@ -5,8 +5,8 @@ homepage: https://github.com/aminsamad/flutter_kurdish_localization
 
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
